fix: export functions for updating multiple labels
