+++ src/app/configurations/crawlconfig/crawlconfig-page.component.ts
@@ -14,10 +14,11 @@ import {DetailDirective} from '../shared/detail.directive';
 import {CrawlConfigDetailsComponent} from './crawlconfig-details/crawlconfig-details.component';
 import {of} from 'rxjs/internal/observable/of';
 import {DeleteDialogComponent} from '../../dialog/delete-dialog/delete-dialog.component';
+import {findLabel, getInitialLabels, intersectLabel, updatedLabels} from '../../commons/group-update/labels/common-labels';
 
 @Component({
   selector: 'app-crawlconfig',
-  template: `    
+  template: `
     <div fxLayout="column" fxLayoutGap="8px">
       <div>
         <app-toolbar>
@@ -151,7 +152,8 @@ export class CrawlConfigPageComponent implements OnInit {
     this.selectedConfigs = crawlConfigs;
     if (!this.singleMode) {
       this.loadComponent(
-        this.mergeCrawlConfigs(crawlConfigs), this.browserConfigs, this.politenessConfigs, getInitialLabels(crawlConfigs),
+        this.mergeCrawlConfigs(crawlConfigs),
+        this.browserConfigs, this.politenessConfigs, getInitialLabels(crawlConfigs, CrawlConfig),
         getInitialExtractText(crawlConfigs), getInitialCreateSnapshot(crawlConfigs), getInitialDepthFirst(crawlConfigs));
     } else {
       this.crawlConfig = crawlConfigs[0];
@@ -420,47 +422,4 @@ function updateMultiDepthFirst(crawlConfig: CrawlConfig, crawlConfigUpdate: Craw
     crawlConfig.depth_first = false;
   }
   return crawlConfig.depth_first;
-}
-
-function getInitialLabels(configs: CrawlConfig[]) {
-  const config = new CrawlConfig();
-  const label = configs.reduce((acc: CrawlConfig, curr: CrawlConfig) => {
-    config.meta.label = intersectLabel(acc.meta.label, curr.meta.label);
-    return config;
-  });
-  return config.meta.label;
-}
-
-
-function findLabel(array: Label[], key, value) {
-  const labelExist = array.find(function (label) {
-    return label.key === key && label.value === value;
-  });
-  if (!labelExist) {
-    return false;
-  }
-  if (labelExist) {
-    return true;
-  }
-}
-
-function intersectLabel(a, b) {
-  const setA = Array.from(new Set(a));
-  const setB = Array.from(new Set(b));
-  const intersection = new Set(setA.filter((x: Label) =>
-    setB.find((label: Label) => x.key === label.key && x.value === label.value) === undefined
-      ? false
-      : true
-  ));
-  return Array.from(intersection);
-}
-
-function updatedLabels(labels) {
-  const result = labels.reduce((unique, o) => {
-    if (!unique.find(obj => obj.key === o.key && obj.value === o.value)) {
-      unique.push(o);
-    }
-    return unique;
-  }, []);
-  return result;
 }
