+++ src/app/configurations/politenessconfig/politenessconfig-page.component.ts
@@ -11,10 +11,11 @@ import {DetailDirective} from '../shared/detail.directive';
 import {FormBuilder} from '@angular/forms';
 import {PolitenessconfigDetailsComponent} from './politenessconfig-details/politenessconfig-details.component';
 import {DeleteDialogComponent} from '../../dialog/delete-dialog/delete-dialog.component';
+import {getInitialLabels, updatedLabels, findLabel, intersectLabel} from '../../commons/group-update/labels/common-labels';
 
 @Component({
   selector: 'app-politenessconfig',
-  template: `    
+  template: `
     <div fxLayout="column" fxLayoutGap="8px">
       <div>
         <app-toolbar>
@@ -128,7 +129,7 @@ export class PolitenessConfigPageComponent implements OnInit {
     this.selectedConfigs = politenessConfigs;
     if (!this.singleMode) {
       this.loadComponent(this.mergePolitenessConfigs(politenessConfigs),
-        this.robotsPolicies, getInitialLabels(politenessConfigs));
+        this.robotsPolicies, getInitialLabels(politenessConfigs, PolitenessConfig));
     } else {
       this.politenessConfig = politenessConfigs[0];
       if (this.componentRef !== null) {
@@ -355,46 +356,4 @@ export class PolitenessConfigPageComponent implements OnInit {
     });
     return config;
   }
-}
-
-function getInitialLabels(configs: PolitenessConfig[]) {
-  const config = new PolitenessConfig();
-  const label = configs.reduce((acc: PolitenessConfig, curr: PolitenessConfig) => {
-    config.meta.label = intersectLabel(acc.meta.label, curr.meta.label);
-    return config;
-  });
-  return config.meta.label;
-}
-
-function findLabel(array: Label[], key, value) {
-  const labelExist = array.find(function (label) {
-    return label.key === key && label.value === value;
-  });
-  if (!labelExist) {
-    return false;
-  }
-  if (labelExist) {
-    return true;
-  }
-}
-
-function intersectLabel(a, b) {
-  const setA = Array.from(new Set(a));
-  const setB = Array.from(new Set(b));
-  const intersection = new Set(setA.filter((x: Label) =>
-    setB.find((label: Label) => x.key === label.key && x.value === label.value) === undefined
-      ? false
-      : true
-  ));
-  return Array.from(intersection);
-}
-
-function updatedLabels(labels) {
-  const result = labels.reduce((unique, o) => {
-    if (!unique.find(obj => obj.key === o.key && obj.value === o.value)) {
-      unique.push(o);
-    }
-    return unique;
-  }, []);
-  return result;
 }
